L02E02.2

+++ Listas/L02E02/Filme.hpp
@@ -16,7 +16,7 @@ class  Filme {
  public :
      Filme(int  id, std::string  nome, std::string  genero, int  duracao);
      ~ Filme();
-     Review * adicionar_review(std::string  comentario, double  nota);
+     Review * adicionar_review(std::string  comentario, float nota);
      void  imprimir_nota_consolidada();
      std::string get_nome();
 };
